fix: show edited values in meal history
